Add support for building with proprietary compiler
Fix incorrect conflict resolution related to SDCLANG

Add Android.bp property sdclang

The sdclang property in Android.bp is the equivalent of LOCAL_SDCLANG in
Android.mk. It can be used in two scenarios.
1) You may set sdclang to false to force individual modules to use AOSP LLVM
even if SDCLANG is set to true globally.
2) You may set sdclang to true to force individual modules to use SDLLVM even if
SDCLANG is set to false globally.

Note: You do not have to set this property if you do not intend to use a target
compiler different from the default one. The default target compiler will be
used if the sdclang property is not set.

Turn off sdclang for cfi sanitizer

This patch turns off sdclang for the cfi sanitizer. This is only a temporary
workaround for the cfi sanitizer issue encountered when building O-MR1 with
SDLLVM. It should be reverted once the cfi sanitizer issue is fixed in SDLLVM.

Revert "Turn off sdclang for cfi sanitizer"

This reverts commit 8ee0c33b1ad5379ada41d80a9926b8e506e9a00b.

Add CFI sanitizer support for SDLLVM

There are two main differences between SDLLVM and AOSP LLVM regarding to using
the CFI sanitizer for Android.
1) SDLLVM uses qcld linker while AOSP LLVM uses gold linker.
2) The flag "-Wl,-plugin-opt,O1" is not included in the CFI LDFLAGS for SDLLVM
compared to what is used by AOSP LLVM.

Change-Id: Ibb4a3548e8cc8ab6c22c745844f3e150b38eb780
Signed-off-by: celtare21 <celtare21@gmail.com>

Add support to enable SDCLANG SA

Change-Id: Iab4be5209de62300ea26ecd74736200978ba9303
Signed-off-by: celtare21 <celtare21@gmail.com>

For aosp build with clang master

Change-Id: I0ecb5027de15cf571497b73aca1b80f3eb069dcc
Signed-off-by: celtare21 <celtare21@gmail.com>